requirements.txt
markdown
django-password-reset
Pillow
django-taggit
django-ckeditor
django-imagekit